third_party/blink/web_tests/svg/dynamic-updates/SVGMarkerElement-dom-orient-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'orient' attribute of the SVGMarkerElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS markerElement.getAttribute('orient') is "45"
PASS markerElement.getAttribute('orient') is "0"
PASS successfullyParsed is true

TEST COMPLETE
